refactor(convex): Align mutations with conventions
Extract syncSupportLastMessage and admin notification preference
helpers so mutation-context callers no longer open extra
transactions via ctx.runMutation; keep the updateLastMessage
internalMutation only for its action caller and delete the two
preference wrappers that had none. Switch the founderWelcome
adminSettings write path from .first() to .unique() to match all
readers of the singleton keys, and add Bounded comments to the
index-scoped and one-time backfill collects.

Resolves: #496
