+++ src/lib/convex/auth.ts
@@ -13,6 +13,10 @@ import authConfig from './auth.config';
 import { requireEnv, googleOAuth, githubOAuth } from './env';
 import { getFounderWelcomeDelay } from './emails/helpers';
 import { incrementCounter } from './admin/counters';
+import {
+	syncAdminPreferences,
+	deactivateAdminPreferencesHelper
+} from './admin/notificationPreferences/helpers';
 import { devNotice } from '../dev/notice';
 
 // Required for triggers to work - references internal auth functions
@@ -143,10 +147,7 @@ export const authComponent = createClient<DataModel, typeof authSchema>(componen
 				// Non-critical: don't block signup if preference creation fails
 				if (user.role === 'admin') {
 					try {
-						await ctx.runMutation(
-							internal.admin.notificationPreferences.mutations.upsertAdminPreferences,
-							{ userId: user._id, email: user.email }
-						);
+						await syncAdminPreferences(ctx, { userId: user._id, email: user.email });
 					} catch (error) {
 						console.error('Failed to create admin preferences:', error);
 					}
@@ -265,22 +266,13 @@ export const authComponent = createClient<DataModel, typeof authSchema>(componen
 				try {
 					if (!wasAdmin && isAdmin) {
 						// Promoted to admin → activate/create preferences
-						await ctx.runMutation(
-							internal.admin.notificationPreferences.mutations.upsertAdminPreferences,
-							{ userId: newUser._id, email: newUser.email }
-						);
+						await syncAdminPreferences(ctx, { userId: newUser._id, email: newUser.email });
 					} else if (wasAdmin && !isAdmin) {
 						// Demoted from admin → deactivate preferences (keep dormant)
-						await ctx.runMutation(
-							internal.admin.notificationPreferences.mutations.deactivateAdminPreferences,
-							{ userId: newUser._id }
-						);
+						await deactivateAdminPreferencesHelper(ctx, newUser._id);
 					} else if (isAdmin && oldUser.email !== newUser.email) {
 						// Admin changed email → update preferences
-						await ctx.runMutation(
-							internal.admin.notificationPreferences.mutations.upsertAdminPreferences,
-							{ userId: newUser._id, email: newUser.email }
-						);
+						await syncAdminPreferences(ctx, { userId: newUser._id, email: newUser.email });
 					}
 				} catch (error) {
 					console.error('Failed to sync admin preferences on user update:', error);
