+++ src/lib/convex/localDev.ts
@@ -1,7 +1,8 @@
 import { v } from 'convex/values';
-import { components, internal } from './_generated/api';
+import { components } from './_generated/api';
 import { internalMutation, type MutationCtx } from './_generated/server';
 import { createAuth } from './auth';
+import { syncAdminPreferences } from './admin/notificationPreferences/helpers';
 
 type BetterAuthUser = {
 	_id: string;
@@ -95,10 +96,7 @@ export const ensureSeededAdmin = internalMutation({
 			});
 		}
 
-		await ctx.runMutation(internal.admin.notificationPreferences.mutations.upsertAdminPreferences, {
-			userId: user._id,
-			email
-		});
+		await syncAdminPreferences(ctx, { userId: user._id, email });
 
 		return {
 			created,
